requirements-travis.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
crispy-semantic-ui==1.0.2
dj-database-url==0.5.0
Django==3.1.1
django-heroku==0.3.1
django-allauth==0.42.0
django-bootstrap4==2.2.0
django-bootstrap-datepicker-plus==3.0.5
django-google-api==1.0.1
django-location-field==2.1.0
django-location-field==2.1.0
django-reversion==3.0.8
django-social-share==1.4.0
geopy==2.0.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
psycopg2==2.7.5
psycopg2-binary==2.8.6
python-dateutil==1.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-heroku==0.3.1
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
oauthlib==3.1.0
psycopg2==2.7.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==1.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
